requirements.txt
langgraph
ipython
typing-extensions
tqdm
 --extra-index-url https://erc.timetoact-group.at/
pydantic
langchain-openai
langchain-core
langchain
pyyaml
erc3>=1.1.0
datasets
tiktoken
nltk
openai
azure-identity
azure-search-documents
pyarrow
llama-index
llama-index-core
python-dotenv